db int to bigint

+++ db/scripts/000001_init_db.up.sql
@@ -31,7 +31,7 @@ CREATE TABLE IF NOT EXISTS roles (
 CREATE TABLE IF NOT EXISTS wallets (
     id bigserial PRIMARY KEY,
     name text NOT NULL,
-    number int UNIQUE NOT NULL CHECK (number >= 111111111111 AND number <= 999999999999),
+    number bigint UNIQUE NOT NULL CHECK (number >= 111111111111 AND number <= 999999999999),
     balance double precision DEFAULT 0 CHECK (balance  >= 0),
     iin text REFERENCES users(iin),
     UNIQUE (number, iin),
